requirements.txt
# Core dependencies for Resume Agent project
anthropic>=0.40.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
PyGithub>=2.1.1

# Data handling
pandas>=2.0.0

# Testing
pytest>=7.4.0
